app opens to match feed if data is in database and repopulates match list from database if there is data. Added an item to options menu to select new event.

+++ app/src/main/java/group8/matchtracker/adapters/PlayerListAdapter.java
@@ -41,7 +41,7 @@ public class PlayerListAdapter extends RecyclerView.Adapter<PlayerListAdapter.Vi
                 Intent i = new Intent(context, TabbedActivity.class);
                 Player selectedPlayer = mPlayers.get(getLayoutPosition());
                 //i.putExtra(DatabaseHelper.EVENT_ID, ((Activity)context).getIntent().getExtras().getInt(DatabaseHelper.EVENT_ID));
-                i.putExtra(DatabaseHelper.TOURNAMENT_ID, ((Activity)context).getIntent().getExtras().getInt(DatabaseHelper.TOURNAMENT_ID));
+                i.putExtra(DatabaseHelper.TOURNAMENT_ID, ((Activity)context).getIntent().getExtras().getLong(DatabaseHelper.TOURNAMENT_ID));
                 i.putExtra(DatabaseHelper.PLAYER_ID, selectedPlayer.getId());
                 context.startActivity(i);
             //}
